requirements.txt
fastapi
uvicorn[standard]
alpaca-trade-api
python-dotenv
pydantic
pyyaml